add authorization and status

+++ backend/controllers/feed.js
@@ -1,5 +1,6 @@
 const {validationResult} = require("express-validator");
 const Post = require('../models/post');
+const User = require('../models/user');
 const {deleteFile} = require("../util/file");
 
 const ITEMS_PER_PAGE = 5;
@@ -9,7 +10,8 @@ exports.getPosts = async (req, res, next) => {
     const page = req.query['page'] || 1;
     const posts = await Post.find()
       .skip((page - 1) * ITEMS_PER_PAGE)
-      .limit(ITEMS_PER_PAGE);
+      .limit(ITEMS_PER_PAGE)
+      .populate('creator');
 
     const totalItems = await Post.countDocuments();
     const totalPages = Math.ceil(totalItems / ITEMS_PER_PAGE);
@@ -44,15 +46,24 @@ exports.postPost = async (req, res, next) => {
       title: req.body.title,
       content: req.body.content,
       imageUrl: req.file.path,
-      creator: {name: 'Fix me'},
+      creator: req.userId,
     })
 
     await post.save();
 
+    const user = await User.findById(req.userId);
+    user.posts.push(post);
+    await user.save();
 
     res.status(201).json({
       message: 'Post created successfully',
-      post
+      post: {
+        ...post._doc,
+        creator: {
+          _id: user._id,
+          name: user.name,
+        }
+      },
     });
   } catch (err) {
     next(err);
@@ -96,6 +107,12 @@ exports.putPost = async (req, res, next) => {
       });
     }
 
+    if (!post.creator.equals(req.userId)) {
+      return res.status(403).json({
+        message: 'You are not allowed to edit this post',
+      });
+    }
+
     post.title = req.body.title;
     post.content = req.body.content;
 
@@ -125,11 +142,53 @@ exports.deletePost = async (req, res, next) => {
       });
     }
 
+    if (!post.creator.equals(req.userId)) {
+      return res.status(403).json({
+        message: 'You are not allowed to delete this post',
+      });
+    }
+
     deleteFile(post.imageUrl).catch(err => console.error(err));
     await Post.deleteOne({_id: req.params['postId']});
 
+    const user = await User.findById(req.userId);
+
+    user.posts.pull(req.params['postId']);
+
+    await user.save();
+
     res.status(200).json({
       message: 'Post deleted successfully',
+    });
+  } catch (err) {
+    next(err);
+  }
+}
+
+exports.getStatus = async (req, res, next) => {
+  try {
+    const user = await User.findById(req.userId);
+
+    res.status(200).json({
+      message: 'Status fetched successfully',
+      status: user.status,
+    });
+  } catch (err) {
+    next(err);
+  }
+}
+
+exports.putStatus = async (req, res, next) => {
+  try {
+    const user = await User.findById(req.userId);
+
+    user.status = req.body.status;
+
+    await user.save();
+
+    res.status(200).json({
+      message: 'Status updated successfully',
+      status: user.status,
     });
   } catch (err) {
     next(err);
